Twitch Censor from channel RebootTech
Hey there, I ran AI Dungeon successfully on Twitch for several weeks.  Here's the censor list that I used.   Bear in mind that you have to censor both the input to the model and the output from the model.   The order is important because the regex considers longer phrases first.